+++ pylearn/model_utils.py
@@ -77,3 +77,51 @@ def load(file_name: str) -> BinaryIO:
     """
     with open(file_name, 'rb') as file:     # rb: reading in binary mode
         return dill.load(file)
+    
+def min_max_normalization(data: np.ndarray) -> np.ndarray:
+    """
+    Normalizes Data in a range from 0 to 1 using minx-max Value of every Feature. 
+    Real-world datasets often contain features that are 
+    varying in degrees of magnitude, range, and units. 
+    Therefore, perform feature scaling, otherwise some features dominate the calculations.
+
+    Parameters:
+        :data (numpy.npdarray): Input for machine learning algorithm
+
+    Returns:
+        Normalized data
+    """
+    #each row is a data point, each column symbolizes a feature
+    #we need to find min and max for every feature
+    min = np.amin(data, axis=0)
+    max = np.amax(data, axis=0)
+    normalized_data = np.ndarray(shape=data.shape)
+
+    for (data_point, feature), value in np.ndenumerate(data):
+        normalized_data[data_point, feature] = (value-min[feature])/(max[feature]-min[feature])
+
+    return normalized_data
+
+def z_normalization(data: np.ndarray) -> np.ndarray:
+    """
+    Normalizes Data in a range from 0 to 1 using Z-Scores.
+    Real-world datasets often contain features that are 
+    varying in degrees of magnitude, range, and units. 
+    Therefore, perform feature scaling, otherwise some features dominate the calculations.
+
+    Parameters:
+        :data (numpy.npdarray): Input for machine learning algorithm
+
+    Returns:
+        Normalized data
+    """
+    #each row is a data point, each column symbolizes a feature
+    #we need to find mean and standart derivation for every feature
+    mean = np.mean(data, axis=0)
+    std = np.std(data, axis=0)
+    normalized_data = np.ndarray(shape=data.shape)
+
+    for (data_point, feature), value in np.ndenumerate(data):
+        normalized_data[data_point, feature] = (value-mean[feature])/(std[feature])
+
+    return normalized_data
